Revert "Merge branch 'master' of https://github.com/nanophotonics/nplab"
This reverts commit 4a54cb9fced9be4a26219dc699a4327e2064c377, reversing
changes made to e7b4bcabcdfa12b4facf83b673ed43828c8beebc.
