+++ nplab/instrument/__init__.py
@@ -34,10 +34,7 @@ class Instrument(ShowGUIMixin):
     """
     __instances = None
     metadata_property_names = () #"Tuple of names of properties that should be automatically saved as HDF5 metadata
-    def __new__(cls, *args, **kwargs):
-        print(f'creating {cls.__name__} instance')
-        return super().__new__(cls)
-    
+
     def __init__(self):
         """Create an instrument object."""
         super(Instrument, self).__init__()
