+++ nplab/instrument/stage/thorlabs_ello/ell9.py
@@ -12,6 +12,12 @@ class Ell9(Ell6):
         return ELL9UI(self)
 
 
+class ELL6UI(QuickControlBox):
+    def __init__(self, instr):
+        super().__init__('ELL6')
+        self.add_spinbox('position', vmin=0, vmax=1)
+        self.auto_connect_by_name(controlled_object=instr)
+
 
 class ELL9UI(QuickControlBox):
     def __init__(self, instr):
